requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
sqlalchemy==2.0.39
alembic==1.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.8.1
jinja2==3.1.5
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
email-validator==2.2.0
